+++ randovania/gui/lib/file_prompts.py
@@ -49,11 +49,12 @@ async def prompt_input_layout(parent: QtWidgets.QWidget) -> Path | None:
     )
 
 
-async def prompt_preset(parent: QtWidgets.QWidget, new_file: bool) -> Path | None:
+async def prompt_preset(parent: QtWidgets.QWidget, new_file: bool, name: str | None = None) -> Path | None:
     """
     Shows an QFileDialog asking the user for a Randovania preset file
     :param parent:
     :param new_file: If it should be an existing file (False) or not.
+    :param name: Default direction for existing file, default output name for new files.
     :return: A string if the user selected a file, None otherwise
     """
     from randovania.layout.versioned_preset import VersionedPreset
@@ -62,4 +63,5 @@ async def prompt_preset(parent: QtWidgets.QWidget, new_file: bool) -> Path | Non
         file_filter=f"Randovania Preset, *.{VersionedPreset.file_extension()};;"
                     f"All Files (*.*)",
         new_file=new_file,
+        current_dir=name,
     )